Main

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

wW :: Float

Largura da janela, em pixeis

wH :: Float

Altura da janela, em pixeis

numeroTiles :: Int

Número de quadrados que preenchem uma linha do jogo

l :: Float

Largura dos quadrados, em proporção com a janela

mapa :: Mapa

Mapa que representa as primeiras cinco linhas (imutáveis) do jogo

data EstadoJogo

Data que define o estado de jogo

Constructors

Menu EstadoJogo EstadoJogo EstadoJogo

Jogar Botoes Numeros Player Fundo FundoAtivo SkinsMapa SkinMapaAtivo SkinsJogador SkinJogadorAtivo

Skins Botoes Numeros Player Fundo FundoAtivo SkinsMapa SkinMapaAtivo SkinsJogador BotaoSkinsSelecionado SkinJogadorAtivo

Opçoes Botoes Numeros Player Fundo FundoAtivo SkinsMapa BotaoOpçoesSelecionado SkinMapaAtivo SkinsJogador SkinJogadorAtivo

Pausa Botoes BotaoPausaSlecionado Numeros Player Fundo FundoAtivo SkinsMapa SkinMapaAtivo SkinsJogador SkinJogadorAtivo

PerdeuJogo Botoes BotaoPerdeuSelecionado Numeros Player Fundo FundoAtivo SkinsMapa SkinMapaAtivo SkinsJogador SkinJogadorAtivo

type Botoes = [Picture]

Imagens Bitmap de todos os botões

type Numeros = [Picture]

Imagens Bitmap de todos os números

data Player

Player (Posição do jogador e matriz do mapa) (Ângulo do jogador) (Acumulador que permite atualizar step) (Lista de números aleatórios) (Score) (Valor que permite mudar a velocidade do ecrã) (HighScore)

Constructors

Player Jogo Float Float [Int] Int Float Int

type Fundo = [Picture]

Imagens Bitmap de todos os fundos

type FundoAtivo = Picture

Fundo que está ativo no jogo

type SkinJogador = [Picture]

Conjunto de Imagens Bitmap de uma só skin

type SkinsJogador = [SkinJogador]

Todas as skins

type SkinJogadorAtivo = [Picture]

Skin ativa

type SkinMapa = [Picture]

Conjunto de Imagens Bitmap de um só mapa

type SkinsMapa = [SkinMapa]

Todos os mapas

type SkinMapaAtivo = [Picture]

Mapa ativo

data BotaoOpçoesSelecionado

Data que define o botão selecionado no menu das opções

Constructors

Mapa1

Mapa2

Mapa3

Mapa4

LightMode

DarkMode

Facil

Medio

Dificil

VoltarOpcao

data BotaoSkinsSelecionado

Data que define o botão selecionado no menu das skins

Constructors

Skin1

Skin2

Skin3

Skin4

Skin5

Skin6

VoltarSkin

data BotaoPausaSlecionado

Data que define o botão selecionado no menu de pausa

Constructors

ReiniciarPausa

MenuPausa

data BotaoPerdeuSelecionado

Data que define o botão selecionado no menu do jogo perdido

Constructors

ReiniciarPerdeu

MenuPerdeu

charToPicture :: Char -> EstadoJogo -> Picture

A função recebe um caractere, que representa um algarismo, e um estado de jogo e, tendo em conta as imagens bitmap presentes na lista de algarismos, devolve a imagem correspondente a esse mesmo algarismo

transformaScore :: String -> EstadoJogo -> Float -> Picture

A função recebe uma string, que representa um número, um estado de jogo e um valor, correspondente ao espaçamento entre algarismos, e devolve a imagem correspondente a esse mesmo número

repeatTiles :: Float -> Picture -> [Picture]

A função recebe um valor, que servirá de acumulador, e uma imagem, i.é., um tile, e devolve uma linha/lista composta pela consecutiva repetição dessa imagem, da esquerda para a direita, ao longo da largura do ecrã

juntaTiles :: Picture -> Picture

A função recebe uma imagem, i.é, um tile e, utilizando a função repeatTiles, transforma a linha/lista numa só imagem

transformaMatriz :: EstadoJogo -> Float -> Picture

Esta função recebe um estado de jogo, que contem um mapa, e um valor, que servirá de acumulador e, utilizando as funções transformaTerreno e transformaObstaculos, desenha todo esse mapa, de cima para baixo, ao longo do ecrã. Nesta função é necessário o pattern matching para os estados de jogo Jogar, Pausa e PerdeuJogo, uma vez que todos estes requerem os desenhos das mesmas imagens

transformaTerreno :: Terreno -> EstadoJogo -> Picture

Esta função recebe um terreno e um estado de jogo e, consoante o tipo de terreno e o mapa ativo, utiliza a função juntaTiles para desenhar uma linha desse terreno

transformaObstaculos :: [Obstaculo] -> Int -> Float -> EstadoJogo -> Picture

Esta função recebe uma lista de obstáculos, um valor que representa a velocidade, um valor que servirá de acumulador, e um estado de jogo e, consoante o tipo de obstáculo e o mapa ativo, desenha uma linha formada pelos obstáculos, da esquerda para a direita, ao longo do ecrã

estendeMapaRecursiva :: Mapa -> [Int] -> Int -> Mapa

Esta função recebe um mapa, uma lista de inteiros e um valor n e, utilizando a função estendeMapa recursivamente, gera e adiciona, ao mapa inicial, n linhas aleatórias

drawWorld :: EstadoJogo -> Picture

Esta função recebe um estado de jogo e, sobrepondo sucessivamente imagens bitmap, devolve o desenho desse respetivo estado. Em alguns estados de jogo, é necessário recorrer ao pattern matching para demonstrar qual o botãoskinmapa que está a ser selecionado

handleEvent :: Event -> EstadoJogo -> EstadoJogo

Esta função recebe um evento e um estado de jogo e, dependendo do tecla pressionada e do estado de jogo em que sem encontra, devolve esse mesmo atualizado ou um outro

step :: Float -> EstadoJogo -> EstadoJogo

Esta função recebe um float e um estado de jogo e, tendo por base as funções deslizeMapa, movimentoEnquantoParado, movimentoObstaculos e jogoTerminou, atualiza o estado de jogo 120 vezes, uma vez que a frame rate escolhida para o jogo é 120

win :: Display

Tipo de display

bkg :: Color

Cor do background

main :: IO ()

A função que roda o jogo