Tarefa1_2022li1g018

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Esta função determina se um dado mapa é válido ou não através de várias condições

group :: Eq a => [a] -> [[a]]

As funções seguintes são algumas funções auxiliares que ajudam na defenição de outras funções. A função group, dada uma lista, agroupa os elementos iguais, ex.: group [1,2,2,3,4,4,4,5,4] -> [[1],[2,2],[3],[4,4,4],[5],[4]]

firsts :: [(Terreno, [Obstaculo])] -> [String]

A função firsts, dada uma lista de tuplos, devolve o primeiro elemento de cada tuplo. Esta função teve de ser mudada de forma a ignorar a velocidade associada aos terrenos Rio e Estrada, devolvendo uma lista de Strings que correspondem ao tipo de terreno (RioEstradaRelva) ex.: firsts [(Rio 3, [Tronco,Tronco,Nenhum]),(Relva, [Arvore,Arvore,Nenhum])] -> [Rio,Relva]

isSuffixOf :: Eq a => [a] -> [a] -> Bool

A função seguinte é uma função auxiliar que ajuda na definição de outras funções. A função isSuffixOf, dada duas listas, determina se a primeira lista é suffixo da segunda ex.: isSuffixOf [20,30] [10,20,30] -> True e isSuffixOf [10,30] [10,20,30] -> False

isPrefixOf :: Eq a => [a] -> [a] -> Bool

A função seguinte é uma função auxiliar que ajuda na definição de outras funções. A função isPrefixOf, dada duas listas, determina se a primeira lista é prefixo da segunda ex.: isPrefixOf [10,20] [10,20,30] -> True e isPrefixOf [10,30] [10,20,30] -> False

terrenoChecker :: Terreno -> [Obstaculo] -> Bool

Esta função determina se, num determinado terreno, certos elementos podem ou não existir. É necessário fazer três casos diferentes uma vez que é a única maneira que temos de distinguir os diferentes tipos de terreno

rioChecker :: [(Terreno, [Obstaculo])] -> Bool

Esta função, dada uma lista de terrenos e os seus obstáculos, de um mapa, verifica se os rios contíguos têm direções opostas. Semenlhantemente à função terrenoChecker, é necessário fazer vários casos dependendo de como se apresentam os terrenos

troncoChecker :: Largura -> [Obstaculo] -> Bool

Esta função, dado um valor de largura e uma lista de obstáculos, verifica se, caso a lista de obstáculos possua troncos, estes não ultrapassam as 5 unidades

troncoCheckerPontas :: [Obstaculo] -> Bool

Esta função, recebe uma lista de obstáculos e verifica se, juntando os troncos das extremidades das listas, se forma um tronco com comprimento superior a 5 unidades

elemTronco :: (Terreno, [Obstaculo]) -> Bool

Esta função, dado um tipo de terreno e uma lista de obstáculos, verifica se, apenas quando o terreno em causa é o rio, poussui na sua lista de obstáculos pelo menos um tronco

carroChecker :: Largura -> [Obstaculo] -> Bool

Esta função, dado um valor de largura e uma lista de obstáculos, verifica se, caso a lista de obstáculos possua carros, estes não ultrapassam as 3 unidades

carroCheckerPontas :: [Obstaculo] -> Bool

Esta função recebe uma lista de obstáculos e verifica se, juntando os Carros das extremidades das listas, se forma um carro com comprimento superior a 3 unidades

velocidadeNulaChecker :: Terreno -> Bool

Esta função, dado um tipo de terreno, verifica se, caso este seja rio ou estrada, possui velocidade não nula associada

terrenoContiguo :: [[String]] -> Bool

Esta função, dada uma lista agrupada (pelo mesmo tipo) de strings (que representam os terrenos), verifica se, atendendo ao tipo de terreno, o número máximo de terrenos contiguos é 4 para os Rios e 5 para as Estradas e Relvas