Tarefa2_2022li1g018

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> [Int] -> Mapa

Esta função recebe um mapa e uma lista de inteiros e, atendendo às funções randomObstaculo e RandomTerreno, gera e adiciona uma nova linha válida ao topo, ou seja, gera um novo mapa

isSuffixOf :: Eq a => [a] -> [a] -> Bool

A função seguinte é uma função auxiliar que ajuda na definição de outras funções. A função isSuffixOf, dada duas listas, determina se a primeira lista é suffixo da segunda ex.: isSuffixOf [20,30] [10,20,30] -> True e isSuffixOf [10,30] [10,20,30] -> False

isPrefixOf :: Eq a => [a] -> [a] -> Bool

A função seguinte é uma função auxiliar que ajuda na definição de outras funções. A função isPrefixOf, dada duas listas, determina se a primeira lista é prefixo da segunda ex.: isPrefixOf [10,20] [10,20,30] -> True e isPrefixOf [10,30] [10,20,30] -> False

proximasVelociadesValidas :: [(Terreno, [Obstaculo])] -> [Int]

A função recebe uma lista de terrenos e os seus obstáculos e, atendendo à velocidade do terreno anterior, devolve uma lista com as velocidades passíveis de serem utilizadas no próximo terreno (rios contíguos não podem ter a mesma direção)

proximosTerrenosValidos :: Mapa -> Int -> [Terreno]

Esta função recebe um Mapa e um número inteiro (que assume o valor de chance) e, consoante a disposição dos tipos de terreno do mapa dado, apresenta uma lista de terrenos, acompanhados pela sua velocidade, que são passíveis de serem procedentes do mapa

randomTerreno :: Mapa -> [Int] -> Terreno

Esta função recebe um Mapa e uma lista de números inteiros e, dependendo do valor dos primeiros dois inteiros da lista e do número de possibilidadesde terrenos válidos, devolve um desses terrenos possíveis

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Esta função recebe uma largura e um terreno e a lista dos seus obstáculos e, atendendo às condições da tarefa 1, devolve uma lista de obstáculos que são passíveis de serem procedentes à lista de obstáculos dada

randomObstaculo :: Int -> (Terreno, [Obstaculo]) -> [Int] -> [Obstaculo]

Esta função recebe uma largura e um terreno e a lista dos seus obstáculos e uma lista de inteiros e, dependendo do valor resultante dos inteiros e do número de possibilidades de obstáculos válidos, devolve uma lista de obstáculos possível para um dado terreno