Tarefa3_2022li1g018

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

Esta função recebe um Jogo, isto é, a posição de um jogador e um mapa válido, e uma jogada. De acordo com a jogada, irá ocorrer uma atualização do mapa bem como da posição do jogador, isto inclui o movimento de obstaculos e do jogador de acordo com o seu ambiente

rotateL :: [Obstaculo] -> Int -> [Obstaculo]

Esta função, dada uma lista de Obstáculos e um número inteiro n (associado à velocidade), devolve a lista deslocada para a esquerda n vezes

rotateR :: [Obstaculo] -> Int -> [Obstaculo]

Esta função, dada uma lista de Obstáculos e um número inteiro n (associado à velocidade), devolve a lista deslocada para a direita n vezes

rotateLestrada :: [Obstaculo] -> [Obstaculo]

Esta função, dada uma lista de Obstáculos, devolve a lista deslocada para a esquerda uma vez

rotateRestrada :: [Obstaculo] -> [Obstaculo]

Esta função, dada uma lista de Obstáculos, devolve a lista deslocada para a direita uma vez

groupObstaculos :: [(Terreno, [Obstaculo])] -> [[Obstaculo]]

Esta função recebe uma lista de terrenos e os seus obstáculos associados e devolve uma lista com todas as listas de obstáculos ex.: groupObstaculos [(Rio 2,[Nenhum,Tronco]),(Relva,[Nenhum,Nenhum])] -> [[Nenhum,Tronco],[Nenhum,Nenhum]]

groupVelocidades :: [(Terreno, [Obstaculo])] -> [Int]

Esta função recebe uma lista de terrenos e os seus obstáculos associados e devolve uma lista com todas as velocidades de cada terreno ex.: groupVelocidades [(Rio 2,[Nenhum,Tronco]),(Relva,[Nenhum,Nenhum])] -> [2,0]

movimentoObstaculos :: Jogador -> [(Terreno, [Obstaculo])] -> Int -> Int -> [(Terreno, [Obstaculo])]

Esta função, dado um jogador, uma lista de terrenos e os seus obstáculos associados, e dois acumuladores inteiros, move os obstáculos para a esquerda ou para a direita, dependendo da sua velocidade, servindo os acumuladores para determinar a posição do jogador (n para a abcissa e k para a ordenada)

moveJogador :: Jogador -> Jogada -> Jogador

Esta função recebe a posição de um jogador e uma jogada, e, dependendo da jogada, move o jogador para uma nova posição

moveJogadorTronco :: Jogador -> Jogada -> [(Terreno, [Obstaculo])] -> Jogador

Esta função recebe a posição de um jogador, uma jogada e uma lista de terrenos e os seus obstáculos associados e, caso o jogador se encontre em cima de um tronco e faça uma jogada, verifica se a jogada ou é para a esquerda ou para a direita, e move o jogador tendo em conta a velocidade com que o tronco se move ex.: (!!) ((!!) (groupObstaculos [(Rio 2, [Nenhum,Tronco]),(Relva, [Nenhum,Nenhum])]) 1) 1 -> -> (!!) ((!!) [[Nenhum,Tronco],[Nenhum,Nenhum]] 1) 1 -> (!!) [Nenhum,Nenhum] 1 -> Nenhum

movimentoEnquantoParado :: Jogador -> [(Terreno, [Obstaculo])] -> Jogador

Esta função recebe uma posição de um jogador e uma lista de terrenos e os seus obstáculos associados e, caso o jogador se encontre na mesma posição que um tronco, este acompanha o movimento do mesmo

barreiraChecker :: Mapa -> Jogador -> Jogada -> Bool

Esta função, dado um estado de um mapa, a posição de um jogador, e uma jogada, determina se, para aquele estado do mapa e para a nova posição desejada pelo jogador, se trata de uma barreira, i.é., um local inacessível. São consideradas barreiras o topo do mapa, os lados esquerdo e direito do mapa e um árvore